+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 All notable changes to this project will be documented in this file.
 ## 0.6.0(unreleased)
 * Refine tests
+* Add section "Limitations" to README.md
 * Add and fix feature/test for tuple as trait field
 
 ## 0.5.0 - March 7, 2024