+++ CHANGELOG.md
@@ -1,12 +1,13 @@
 # Changelog
 
 All notable changes to this project will be documented in this file.
-## 0.6.0(unreleased)
+## 0.6.0- March 9, 2024
 * Refine tests
 * Add section "Limitations" to README.md
 * Add and fix feature/test for tuple as trait field
 * Add feature/test for HashSet as trait field
 * Add feature/test for Customized Struct as trait field
+* Add feature/test for Enum as trait field
 * Add feature/test for Dict(BtreeMap, similar to HashMap, but in order) as trait field
 
 ## 0.5.0 - March 7, 2024